+++ CHANGELOG.md
@@ -11,6 +11,7 @@ v0.31.1 (master)
 * Add a ContactBlogConfig widget to notify the user about a Private blog configuration
 * Fix #1270 Add PeerTube embedding support in Posts
 * Refactor the embbeding code and optimize the eager loading of attachments in Posts
+* Add a "Scroll to Message in History" feature
 
 v0.31
 ---------------------------